+++ designer/DesignerApp.ts
@@ -42,7 +42,7 @@ class DesignerApp {
 
   private _onGrammarKeyup() {
     this._grammarDidUpdate()
-    this._codeDidUpdate()
+    this._onCodeKeyUp()
     // Hack to break CM cache:
     if (true) {
       const val = this.codeInstance.getValue()
@@ -71,7 +71,7 @@ class DesignerApp {
       .register()
       .fromTextAreaWithAutocomplete(<any>this._codeConsole[0], { lineWrapping: true })
 
-    this.codeInstance.on("keyup", () => this._codeDidUpdate())
+    this.codeInstance.on("keyup", () => this._onCodeKeyUp())
 
     // loadFromURL
     const wasLoadedFromDeepLink = await this._loadFromDeepLink()
@@ -94,6 +94,7 @@ class DesignerApp {
   private _execResultsTextArea = jQuery("#execResultsTextArea")
   private _compileButton = jQuery("#compileButton")
   private _explainButton = jQuery("#explainButton")
+  private _explainRootsButton = jQuery("#explainRootsButton")
   private _downloadButton = jQuery("#downloadButton")
   private _samplesButtons = jQuery("#samplesButtons")
   private _otherErrorsDiv = jQuery("#otherErrorsDiv")
@@ -113,17 +114,22 @@ class DesignerApp {
   }
 
   private _bindListeners() {
+    const that = this
     this._execButton.on("click", () => {
       this._setProgramResults(this.program ? this.program.executeSync() : "Program failed to execute")
     })
     this._compileButton.on("click", () => {
       this._setProgramResults(this.program ? this.program.compile() : "Program failed to execute")
     })
+
+    this._explainRootsButton.on("click", () => {
+      this._setProgramResults(this.program ? this.program.getInPlacePreludeCellTypeTreeWithNodeConstructorNames() : "Program failed to parse")
+    })
+
     this._explainButton.on("click", () => {
       this._setProgramResults(this.program ? this.program.getInPlaceCellTypeTreeWithNodeConstructorNames() : "Program failed to parse")
     })
 
-    const that = this
     this._samplesButtons.on("click", "a", function() {
       if (jQuery(this).hasClass("resetButton")) {
         that.resetCommand()
@@ -145,7 +151,13 @@ class DesignerApp {
 
     this._simulateDataButton.on("click", () => {
       const grammarProgram = new jtree.GrammarProgram(this.grammarInstance.getValue())
-      this.codeInstance.setValue(grammarProgram.generateSimulatedData())
+      this.codeInstance.setValue(
+        grammarProgram
+          ._getRootNodeTypeDefinitionNode()
+          .generateSimulatedData()
+          .join("\n")
+      )
+      this._onCodeKeyUp()
     })
 
     this._downloadButton.on("click", () => this._downloadBundleCommand())
@@ -245,7 +257,7 @@ class DesignerApp {
     return "#" + encodeURIComponent(tree.toString())
   }
 
-  private _codeDidUpdate() {
+  private _onCodeKeyUp() {
     const code = this.codeInstance.getValue()
     this._updateLocalStorage()
     const programConstructor = this._getGrammarConstructor()
@@ -268,7 +280,7 @@ class DesignerApp {
         .forEach((err: any) => {
           const el = err.getCodeMirrorLineWidgetElement(() => {
             this.codeInstance.setValue(this.program.toString())
-            this._codeDidUpdate()
+            this._onCodeKeyUp()
           })
           this.codeWidgets.push(this.codeInstance.addLineWidget(err.getLineNumber() - 1, el, { coverGutter: false, noHScroll: false }))
         })
@@ -283,7 +295,7 @@ class DesignerApp {
     this.codeInstance.setValue(code)
     this._clearHash()
     this._grammarDidUpdate()
-    this._codeDidUpdate()
+    this._onCodeKeyUp()
     this._clearResults()
   }