+++ products/jtree.browser.js
@@ -2652,15 +2652,6 @@ class GrammarBackedNode extends TreeNode {
   _getErrorNodeErrors() {
     return [this.getFirstWord() ? new UnknownNodeTypeError(this) : new BlankLineError(this)]
   }
-  generateSimulatedData(nodeCount = 1) {
-    const lines = []
-    const cells = this._getGrammarBackedCellArray()
-    while (nodeCount > 0) {
-      lines.push(cells.map(cell => cell.generateSimulatedData()).join(" "))
-      nodeCount--
-    }
-    return lines.join("\n")
-  }
   _getBlobNodeCatchAllNodeType() {
     return BlobNode
   }
@@ -2863,6 +2854,11 @@ class GrammarBackedRootNode extends GrammarBackedNode {
       .map(child => child.constructor.name + this.getZI() + child.getIndentation() + child.getLineCellTypes())
       .join("\n")
   }
+  getInPlacePreludeCellTypeTreeWithNodeConstructorNames() {
+    return this.getTopDownArray()
+      .map(child => child.constructor.name + this.getZI() + child.getIndentation() + child.getLineCellPreludeTypes())
+      .join("\n")
+  }
   getTreeWithNodeTypes() {
     return this.getTopDownArray()
       .map(child => child.constructor.name + this.getZI() + child.getIndentation() + child.getLine())
@@ -2908,36 +2904,17 @@ class GrammarBackedNonRootNode extends GrammarBackedNode {
     return this._getGrammarBackedCellArray().filter(cell => cell.getWord() !== undefined)
   }
   _getGrammarBackedCellArray() {
-    const definition = this.getDefinition()
-    const grammarProgram = definition.getLanguageDefinitionProgram()
-    const requiredCellTypeIds = definition.getRequiredCellTypeIds()
-    const numberOfRequiredCells = requiredCellTypeIds.length
-    const catchAllCellTypeId = definition.getCatchAllCellTypeId()
-    const actualWordCountOrRequiredCellCount = Math.max(this.getWords().length, numberOfRequiredCells)
-    const cells = []
-    // A for loop instead of map because "numberOfCellsToFill" can be longer than words.length
-    for (let cellIndex = 0; cellIndex < actualWordCountOrRequiredCellCount; cellIndex++) {
-      const isCatchAll = cellIndex >= numberOfRequiredCells
-      let cellTypeId
-      if (isCatchAll) cellTypeId = catchAllCellTypeId
-      else cellTypeId = requiredCellTypeIds[cellIndex]
-      let cellTypeDefinition = grammarProgram.getCellTypeDefinitionById(cellTypeId)
-      let cellConstructor
-      if (cellTypeDefinition) cellConstructor = cellTypeDefinition.getCellConstructor()
-      else if (cellTypeId) cellConstructor = GrammarUnknownCellTypeCell
-      else {
-        cellConstructor = GrammarExtraWordCellTypeCell
-        cellTypeId = PreludeCellTypeIds.extraWordCell
-        cellTypeDefinition = grammarProgram.getCellTypeDefinitionById(cellTypeId)
-      }
-      cells[cellIndex] = new cellConstructor(this, cellIndex, cellTypeDefinition, cellTypeId, isCatchAll)
-    }
-    return cells
+    return this.getDefinition()._getGrammarBackedCellArray(this, this.getWords())
   }
   // todo: just make a fn that computes proper spacing and then is given a node to print
   getLineCellTypes() {
     return this._getGrammarBackedCellArray()
       .map(slot => slot.getCellTypeId())
+      .join(" ")
+  }
+  getLineCellPreludeTypes() {
+    return this._getGrammarBackedCellArray()
+      .map(slot => slot._getCellTypeDefinition()._getPreludeKindId())
       .join(" ")
   }
   getLineHighlightScopes(defaultScope = "source") {
@@ -3017,13 +2994,16 @@ class UnknownNodeTypeNode extends GrammarBackedNonRootNode {
 A cell contains a word but also the type information for that word.
 */
 class AbstractGrammarBackedCell {
-  constructor(node, index, typeDef, cellTypeId, isCatchAll) {
+  constructor(node, index, typeDef, cellTypeId, isCatchAll, nodeTypeDef) {
     this._typeDef = typeDef
     this._node = node
     this._isCatchAll = isCatchAll
     this._index = index
     this._cellTypeId = cellTypeId
-    this._word = node.getWord(index)
+    this._nodeTypeDefinition = nodeTypeDef
+  }
+  getWord() {
+    return this._node.getWord(this._index)
   }
   getCellTypeId() {
     return this._cellTypeId
@@ -3054,8 +3034,11 @@ class AbstractGrammarBackedCell {
       }
     })
   }
-  getWord() {
-    return this._word
+  generateSimulatedDataForCell() {
+    // todo: cleanup
+    const cellDef = this._getCellTypeDefinition()
+    const enumOptions = cellDef._getFromExtended(GrammarConstants.enum)
+    return enumOptions ? TreeUtils.getRandomString(1, enumOptions.split(" ")) : this._generateSimulatedDataForCell()
   }
   _getCellTypeDefinition() {
     return this._typeDef
@@ -3068,63 +3051,70 @@ class AbstractGrammarBackedCell {
   }
   isValid() {
     const runTimeOptions = this.getNode().getRunTimeEnumOptions(this)
-    if (runTimeOptions) return runTimeOptions.includes(this._word)
-    return this._getCellTypeDefinition().isValid(this._word, this.getNode().getRootProgramNode()) && this._isValid()
+    const word = this.getWord()
+    if (runTimeOptions) return runTimeOptions.includes(word)
+    return this._getCellTypeDefinition().isValid(word, this.getNode().getRootProgramNode()) && this._isValid()
   }
   getErrorIfAny() {
-    if (this._word !== undefined && this.isValid()) return undefined
+    const word = this.getWord()
+    if (word !== undefined && this.isValid()) return undefined
     // todo: refactor invalidwordError. We want better error messages.
-    return this._word === undefined || this._word === "" ? new MissingWordError(this) : new InvalidWordError(this)
+    return word === undefined || word === "" ? new MissingWordError(this) : new InvalidWordError(this)
   }
 }
 AbstractGrammarBackedCell.parserFunctionName = ""
 class GrammarIntCell extends AbstractGrammarBackedCell {
   _isValid() {
-    const num = parseInt(this._word)
+    const word = this.getWord()
+    const num = parseInt(word)
     if (isNaN(num)) return false
-    return num.toString() === this._word
+    return num.toString() === word
   }
-  generateSimulatedData() {
+  _generateSimulatedDataForCell() {
     return TreeUtils.getRandomString(2, "0123456789".split(""))
   }
   getRegexString() {
     return "-?[0-9]+"
   }
   getParsed() {
-    return parseInt(this._word)
+    const word = this.getWord()
+    return parseInt(word)
   }
 }
 GrammarIntCell.defaultHighlightScope = "constant.numeric.integer"
 GrammarIntCell.parserFunctionName = "parseInt"
 class GrammarBitCell extends AbstractGrammarBackedCell {
   _isValid() {
-    const str = this._word
-    return str === "0" || str === "1"
+    const word = this.getWord()
+    return word === "0" || word === "1"
   }
-  generateSimulatedData() {
+  _generateSimulatedDataForCell() {
     return TreeUtils.getRandomString(1, "01".split(""))
   }
   getRegexString() {
     return "[01]"
   }
   getParsed() {
-    return !!parseInt(this._word)
+    const word = this.getWord()
+    return !!parseInt(word)
   }
 }
 GrammarBitCell.defaultHighlightScope = "constant.numeric"
 class GrammarFloatCell extends AbstractGrammarBackedCell {
   _isValid() {
-    const num = parseFloat(this._word)
-    return !isNaN(num) && /^-?\d*(\.\d+)?$/.test(this._word)
+    const word = this.getWord()
+    const num = parseFloat(word)
+    return !isNaN(num) && /^-?\d*(\.\d+)?$/.test(word)
   }
-  generateSimulatedData() {
+  _generateSimulatedDataForCell() {
     return TreeUtils.getRandomString(2, "0123456789".split("")) + "." + TreeUtils.getRandomString(2, "0123456789".split(""))
   }
   getRegexString() {
     return "-?d*(.d+)?"
   }
   getParsed() {
-    return parseFloat(this._word)
+    const word = this.getWord()
+    return parseFloat(word)
   }
 }
 GrammarFloatCell.defaultHighlightScope = "constant.numeric.float"
@@ -3137,10 +3127,11 @@ class GrammarBoolCell extends AbstractGrammarBackedCell {
     this._falses = new Set(["0", "false", "f", "no"])
   }
   _isValid() {
-    const str = this._word.toLowerCase()
+    const word = this.getWord()
+    const str = word.toLowerCase()
     return this._trues.has(str) || this._falses.has(str)
   }
-  generateSimulatedData() {
+  _generateSimulatedDataForCell() {
     return TreeUtils.getRandomString(1, ["1", "true", "t", "yes", "0", "false", "f", "no"])
   }
   _getOptions() {
@@ -3150,7 +3141,8 @@ class GrammarBoolCell extends AbstractGrammarBackedCell {
     return "(?:" + this._getOptions().join("|") + ")"
   }
   getParsed() {
-    return this._trues.has(this._word.toLowerCase())
+    const word = this.getWord()
+    return this._trues.has(word.toLowerCase())
   }
 }
 GrammarBoolCell.defaultHighlightScope = "constant.numeric"
@@ -3158,27 +3150,31 @@ class GrammarAnyCell extends AbstractGrammarBackedCell {
   _isValid() {
     return true
   }
-  generateSimulatedData() {
-    return TreeUtils.getRandomString(10)
+  _generateSimulatedDataForCell() {
+    return TreeUtils.getRandomString(1, "two roads diverged in a yellow wood".split(" "))
   }
   getRegexString() {
     return "[^ ]+"
   }
   getParsed() {
-    return this._word
+    return this.getWord()
   }
 }
-class GrammarKeywordCell extends GrammarAnyCell {}
+class GrammarKeywordCell extends GrammarAnyCell {
+  _generateSimulatedDataForCell() {
+    return this._nodeTypeDefinition._getId().replace(GrammarProgram.nodeTypeSuffixRegex, "")
+  }
+}
 GrammarKeywordCell.defaultHighlightScope = "keyword"
 class GrammarExtraWordCellTypeCell extends AbstractGrammarBackedCell {
   _isValid() {
     return false
   }
-  generateSimulatedData() {
-    return ""
+  _generateSimulatedDataForCell() {
+    return "extraWord" // should never occur?
   }
   getParsed() {
-    return this._word
+    return this.getWord()
   }
   getErrorIfAny() {
     return new ExtraWordError(this)
@@ -3188,11 +3184,11 @@ class GrammarUnknownCellTypeCell extends AbstractGrammarBackedCell {
   _isValid() {
     return false
   }
-  generateSimulatedData() {
-    return ""
+  _generateSimulatedDataForCell() {
+    return "unknownCell" // should never occur?
   }
   getParsed() {
-    return this._word
+    return this.getWord()
   }
   getErrorIfAny() {
     return new UnknownCellTypeError(this)
@@ -3517,18 +3513,16 @@ class cellTypeDefinitionNode extends AbstractExtendibleTreeNode {
     return this._getPreludeKind() || GrammarAnyCell
   }
   _getPreludeKind() {
-    const kinds = {}
-    kinds[PreludeCellTypeIds.anyCell] = GrammarAnyCell
-    kinds[PreludeCellTypeIds.keywordCell] = GrammarKeywordCell
-    kinds[PreludeCellTypeIds.floatCell] = GrammarFloatCell
-    kinds[PreludeCellTypeIds.numberCell] = GrammarFloatCell
-    kinds[PreludeCellTypeIds.bitCell] = GrammarBitCell
-    kinds[PreludeCellTypeIds.boolCell] = GrammarBoolCell
-    kinds[PreludeCellTypeIds.intCell] = GrammarIntCell
-    return kinds[this.getWord(0)] || kinds[this._getExtendedCellTypeId()]
+    return PreludeKinds[this.getWord(0)] || PreludeKinds[this._getExtendedCellTypeId()]
+  }
+  _getPreludeKindId() {
+    if (PreludeKinds[this.getWord(0)]) return this.getWord(0)
+    else if (PreludeKinds[this._getExtendedCellTypeId()]) return this._getExtendedCellTypeId()
+    return PreludeCellTypeIds.anyCell
   }
   _getExtendedCellTypeId() {
-    return this.get(GrammarConstants.extends)
+    const arr = this._getAncestorsArray()
+    return arr[arr.length - 1]._getId()
   }
   getHighlightScope() {
     const hs = this._getFromExtended(GrammarConstants.highlightScope)
@@ -3966,6 +3960,54 @@ ${captures}
     const ancestorIds = this.getAncestorNodeTypeIdsArray()
     if (ancestorIds.length > 1) return ancestorIds[ancestorIds.length - 2]
   }
+  generateSimulatedData(nodeCount = 1, indentCount = -1, nodeTypeChain = []) {
+    const nodeTypeIds = this._getInScopeNodeTypeIds()
+    const catchAllNodeTypeId = this._getFromExtended(GrammarConstants.catchAllNodeType)
+    if (catchAllNodeTypeId) nodeTypeIds.push(catchAllNodeTypeId)
+    const thisId = this._getId()
+    const lines = []
+    const cells = this._getGrammarBackedCellArray()
+    while (nodeCount > 0) {
+      if (cells.length) lines.push(" ".repeat(indentCount >= 0 ? indentCount : 0) + cells.map(cell => cell.generateSimulatedDataForCell()).join(" "))
+      nodeTypeIds.forEach(nodeTypeId => {
+        if (nodeTypeChain.indexOf(nodeTypeId) > -1) return true
+        const def = this.getNodeTypeDefinitionByNodeTypeId(nodeTypeId)
+        const chain = nodeTypeChain.slice(0)
+        chain.push(nodeTypeId)
+        def.generateSimulatedData(1, indentCount + 1, chain).forEach(line => {
+          lines.push(line)
+        })
+      })
+      nodeCount--
+    }
+    return lines
+  }
+  _getGrammarBackedCellArray(node = undefined, words = []) {
+    const grammarProgram = this.getLanguageDefinitionProgram()
+    const requiredCellTypeIds = this.getRequiredCellTypeIds()
+    const numberOfRequiredCells = requiredCellTypeIds.length
+    const catchAllCellTypeId = this.getCatchAllCellTypeId()
+    const actualWordCountOrRequiredCellCount = Math.max(words.length, numberOfRequiredCells)
+    const cells = []
+    // A for loop instead of map because "numberOfCellsToFill" can be longer than words.length
+    for (let cellIndex = 0; cellIndex < actualWordCountOrRequiredCellCount; cellIndex++) {
+      const isCatchAll = cellIndex >= numberOfRequiredCells
+      let cellTypeId
+      if (isCatchAll) cellTypeId = catchAllCellTypeId
+      else cellTypeId = requiredCellTypeIds[cellIndex]
+      let cellTypeDefinition = grammarProgram.getCellTypeDefinitionById(cellTypeId)
+      let cellConstructor
+      if (cellTypeDefinition) cellConstructor = cellTypeDefinition.getCellConstructor()
+      else if (cellTypeId) cellConstructor = GrammarUnknownCellTypeCell
+      else {
+        cellConstructor = GrammarExtraWordCellTypeCell
+        cellTypeId = PreludeCellTypeIds.extraWordCell
+        cellTypeDefinition = grammarProgram.getCellTypeDefinitionById(cellTypeId)
+      }
+      cells[cellIndex] = new cellConstructor(node, cellIndex, cellTypeDefinition, cellTypeId, isCatchAll, this)
+    }
+    return cells
+  }
 }
 // todo: remove?
 class nodeTypeDefinitionNode extends AbstractGrammarDefinitionNode {}
@@ -4298,6 +4340,14 @@ GrammarProgram.cellTypeSuffixRegex = new RegExp(GrammarConstants.cellTypeSuffix 
 GrammarProgram.cellTypeFullRegex = new RegExp("^[a-zA-Z0-9_]+" + GrammarConstants.cellTypeSuffix + "$")
 GrammarProgram._languages = {}
 GrammarProgram._nodeTypes = {}
+const PreludeKinds = {}
+PreludeKinds[PreludeCellTypeIds.anyCell] = GrammarAnyCell
+PreludeKinds[PreludeCellTypeIds.keywordCell] = GrammarKeywordCell
+PreludeKinds[PreludeCellTypeIds.floatCell] = GrammarFloatCell
+PreludeKinds[PreludeCellTypeIds.numberCell] = GrammarFloatCell
+PreludeKinds[PreludeCellTypeIds.bitCell] = GrammarBitCell
+PreludeKinds[PreludeCellTypeIds.boolCell] = GrammarBoolCell
+PreludeKinds[PreludeCellTypeIds.intCell] = GrammarIntCell
 window.GrammarConstants = GrammarConstants
 window.PreludeCellTypeIds = PreludeCellTypeIds
 window.GrammarProgram = GrammarProgram